+++ contracts/azorius/BaseStrategy.sol
@@ -43,7 +43,7 @@ abstract contract BaseStrategy is OwnableUpgradeable, FactoryFriendly, IBaseStra
     function isProposer(address _address) external view virtual returns (bool);
 
     /** @inheritdoc IBaseStrategy*/
-    function votingEndBlock(uint256 _proposalId) external view virtual returns (uint64);
+    function votingEndBlock(uint256 _proposalId) external view virtual returns (uint32);
 
     /**
      * Sets the address of the Azorius module contract.
